requirements.txt
# ─── LlamaIndex ───
llama-index-core>=0.14.0
llama-index-llms-ollama>=0.9.0
llama-index-embeddings-huggingface>=0.6.0
llama-index-readers-file>=0.5.0

# ─── LangChain ───
langchain>=1.0.0
langchain-ollama>=1.0.0
langchain-community>=0.4.0

# ─── Vektör DB & Embedding ───
chromadb>=1.0.0
sentence-transformers>=5.0.0

# ─── Web Arayüzü ───
flask>=3.0.0

# ─── Yardımcı ───
tiktoken>=0.8.0
pypdf>=5.0.0
psutil>=6.0.0
